+++ manager-ui/app/pages/my-space/domains/[domainId]/recipients/create.vue
@@ -1,8 +1,6 @@
 <script setup lang="ts">
 definePageMeta({});
 
-const MB = 1024 * 1024;
-
 const { t } = useI18n();
 const { call } = useApi();
 const { apiErrorMessage } = useApiError();